WIP: Search Screen Overhaul Added title type selection to DiscoverScreen.kt and introduced searchTitleType stateflow to DiscoverViewModel.kt in order to have a one common searchCriteria stateflow that determines which search with which query is being requested from the api.
DiscoverScreen.kt now displays a list with results once the search keyboard button is clicked.

Adjusted TitlesRemoteMediatorProviderImpl.kt to initialize REFRESH if the query is the same but request type has changed in order to get new results if query remained the same but a different list is requested from the api.

Removed Loading and Error states from SearchViewState.kt as those are handled by the internal pager's load state.

Made changed to RecentSearchesDao.kt and TitlesLocalDataSource.kt so that first we check if the query was already searched and if so - overwrite the previous entry with new search date, and if not - insert as new entry.

+++ app/src/main/java/com/myapplications/mywatchlist/data/local/titles/TitlesLocalDataSource.kt
@@ -210,14 +210,21 @@ class TitlesLocalDataSourceImpl @Inject constructor(
             .map { it.toRecentSearchList() }
 
     override suspend fun saveNewRecentSearch(newSearch: String) = withContext(dispatcher) {
-        if (recentSearchesDao.getCountOfEntries() >= 100) {
-            recentSearchesDao.deleteOldestRecentSearch()
-        }
-        recentSearchesDao.insertRecentSearch(
-            RecentSearchEntity(
-                searchedString = newSearch,
-                searchedDateTime = LocalDateTime.now()
+        if (recentSearchesDao.checkIfSearchExists(newSearch)) {
+            val existingRecentSearch = recentSearchesDao.getRecentSearchBySearchText(newSearch)
+            recentSearchesDao.insertRecentSearch(
+                recentSearchEntity = existingRecentSearch.copy(searchedDateTime = LocalDateTime.now())
             )
-        )
+        } else {
+            if (recentSearchesDao.getCountOfEntries() >= 100) {
+                recentSearchesDao.deleteOldestRecentSearch()
+            }
+            recentSearchesDao.insertRecentSearch(
+                RecentSearchEntity(
+                    searchedString = newSearch,
+                    searchedDateTime = LocalDateTime.now()
+                )
+            )
+        }
     }
 }
